Split FMonad instances using hs-boot
