+++ functor-monad/src/FMonad.hs-boot
@@ -1,0 +1,27 @@
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE QuantifiedConstraints #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeOperators #-}
+module FMonad
+  ( type (~>),
+    FFunctor (..),
+    FMonad (..),
+    fjoin
+  )
+where
+
+import {-# SOURCE #-} FFunctor
+
+import Data.Functor.Compose ( Compose )
+import Data.Functor.Product ( Product )
+import Data.Functor.Sum ( Sum )
+
+class FFunctor ff => FMonad ff where
+  fpure :: (Functor g) => g ~> ff g
+  fbind :: (Functor g, Functor h) => (g ~> ff h) -> ff g a -> ff h a
+
+fjoin :: (FMonad ff, Functor g) => ff (ff g) ~> ff g
+
+instance Functor f => FMonad (Sum f)
+instance (Functor f, forall a. Monoid (f a)) => FMonad (Product f)
+instance Monad f => FMonad (Compose f)
